Installation Page code implemented following a better structure
