+++ setup/eatup_setup_front/src/model/Callbacks.ts
@@ -1,0 +1,5 @@
+interface InstallationCallback {
+    (dbUsr: string, dbUsrPasswd: string, serverPort: number): void;
+}
+
+export default InstallationCallback;
