+++ setup/eatup_setup_front/src/pages/Installation/Ids.ts
@@ -1,0 +1,7 @@
+enum Ids {
+    dbUsr = "dbUsr",
+    dbPasswd = "dbPasswd",
+    serverPort = "serverPort"
+};
+
+export default Ids;
